requirements.txt
setuptools==39.1.0
easydict==1.9
seaborn==0.9.0
opencv_python==4.0.0.21
requests==2.21.0
torch==1.0.1.post2
torchvision==0.2.2
prettytable==0.7.2
numpy==1.16.2
tqdm==4.29.1
coloredlogs==10.0
scipy==1.2.1
setproctitle==1.1.10
tensorflow_gpu==1.9.0
matplotlib==2.0.2
visdom==0.1.8.8
Cython==0.29.6
Pillow==6.1
tensorboard
progress
h5py
six
tb-nightly
future
yacs
gdown
flake8
yapf
isort